@- Changed WDDX to always serialize arrays as structs. (Andrei) Changed WDDX to always serialize arrays as structs. PR #4924
git-svn-id: http://svn.php.net/repository/php/php-src/php/php-src/trunk@25863 c90b9560-bf6c-de11-be94-00142212c4b1

+++ main/fopen_wrappers.c
@@ -987,7 +987,7 @@ static FILE *php_fopen_url_wrapper(const char *path, char *mode, int options, in
 	const char *protocol=NULL;
 	int n=0;
 
-	for(p=path;isalnum(*p);p++) 
+	for(p=path;isalnum((int)*p);p++) 
 		n++;
 	if((*p==':')&&(n>1)) {
 		protocol=path;
